Created game list, modified navigation and renamed package
