+++ app/src/main/AndroidManifest.xml
@@ -22,10 +22,11 @@
         </activity>
         <activity android:name=".game.QuestionActivity">
         </activity>
-        <activity android:name=".match.PlayerListAdminActivity">
+        <activity android:name=".join.PlayerListAdminActivity">
         </activity>
-        <activity android:name=".match.PlayerListGuestActivity">
-
+        <activity android:name=".join.PlayerListGuestActivity">
+        </activity>
+        <activity android:name=".join.GamesListActivity">
         </activity>
     </application>
