+++ app/src/main/res/drawable/ic_play.xml
@@ -1,0 +1,48 @@
+<vector android:height="24dp"
+        android:viewportHeight="268.856"
+        android:viewportWidth="268.856"
+        android:width="24dp"
+        xmlns:android="http://schemas.android.com/apk/res/android">
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M46.14,233.85c-1.22,-3.67 -6.12,-3.67 -7.34,0c-2.45,8.57 -0.61,19.58 1.22,28.15c0.61,2.45 4.9,2.45 5.51,0C46.75,253.44 48.59,242.42 46.14,233.85z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M45.53,187.34c-0.61,-3.06 -5.51,-3.06 -6.12,0c-1.84,7.34 0.61,16.52 1.22,23.87c0,2.45 3.67,2.45 3.67,0C44.92,203.86 47.36,195.3 45.53,187.34z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M43.69,133.48c-6.12,7.34 -4.28,20.19 -3.67,28.76c0,2.45 4.28,2.45 4.28,0c0.61,-4.9 0.61,-9.79 1.22,-14.69c0.61,-4.28 1.84,-7.95 2.45,-12.24C48.59,132.87 45.53,131.65 43.69,133.48z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M48.59,74.73c-0.61,-2.45 -4.28,-2.45 -5.51,0c-1.84,4.9 -0.61,9.79 -0.61,14.69c0,5.51 0.61,11.02 1.22,15.91c0,3.06 4.28,3.06 4.28,0c0.61,-5.51 0.61,-11.02 1.22,-15.91C49.2,84.52 49.81,79.02 48.59,74.73z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M46.75,36.18c0.61,-3.67 1.84,-7.34 2.45,-11.02c0,-2.45 -3.06,-3.67 -4.9,-1.84c-6.73,6.12 -3.67,19.58 0.61,26.32c1.22,1.84 4.28,0.61 3.67,-1.22C46.75,44.13 46.14,40.46 46.75,36.18z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M81.64,14.15c-6.12,-9.18 -22.03,-11.63 -31.82,-14.08c-2.45,-0.61 -4.28,3.06 -1.84,4.28c4.9,1.84 9.79,3.67 14.69,6.12c5.51,2.45 9.79,5.51 14.69,7.96C80.41,19.65 83.47,16.59 81.64,14.15z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M126.31,35.56c-6.73,-1.84 -12.85,-6.73 -19.58,-9.18c-3.06,-1.22 -6.12,3.06 -3.06,4.9c3.67,2.45 7.34,4.9 11.02,7.34s6.73,3.67 11.63,3.06C129.98,41.07 129.37,36.18 126.31,35.56z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M175.27,73.51c-3.06,-4.28 -7.96,-6.12 -11.63,-9.18c-4.28,-3.06 -8.57,-6.73 -12.85,-10.4c-2.45,-1.84 -4.9,1.22 -3.06,3.06c3.67,4.28 6.73,7.96 11.02,11.63c3.67,3.67 7.96,8.57 13.46,9.79C174.05,79.02 176.5,75.96 175.27,73.51z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M218.11,97.38c-2.45,-3.06 -6.12,-4.28 -9.18,-5.51c-3.67,-1.84 -7.96,-3.67 -11.63,-5.51c-3.06,-1.22 -6.12,3.06 -3.06,4.9c3.67,1.84 7.34,4.28 11.02,6.12c3.06,1.84 6.73,4.28 10.4,4.28C217.5,101.66 219.34,99.21 218.11,97.38z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M226.68,116.35c-4.28,3.67 -7.96,7.96 -11.63,12.85c-2.45,3.06 -4.9,6.73 -2.45,11.02c1.23,1.84 3.67,2.45 4.9,0.61c2.45,-2.45 3.06,-4.9 4.28,-7.34c2.45,-4.28 5.51,-8.57 8.57,-12.85C232.8,117.57 229.74,113.9 226.68,116.35z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M193.63,161.02c-1.84,2.45 -3.06,4.28 -5.51,6.12c-1.23,1.23 -4.28,3.67 -6.12,3.67c-3.67,0 -5.51,3.67 -3.67,6.73c1.84,3.06 6.12,2.45 7.96,0c0.61,-1.22 2.45,-1.84 3.67,-3.06c1.22,-1.23 2.45,-1.84 3.06,-3.06c1.84,-2.45 3.67,-6.12 4.28,-9.18C197.92,160.41 194.85,159.19 193.63,161.02z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M163.03,197.13c-2.45,3.06 -4.9,6.12 -7.95,9.18c-3.06,2.45 -7.34,3.67 -9.18,6.73c-1.84,3.06 0,5.51 3.06,5.51c4.28,0.61 8.57,-4.9 11.63,-7.34c3.06,-3.06 6.12,-7.34 7.96,-11.63C169.76,196.52 164.87,193.46 163.03,197.13z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M125.09,228.96c-3.06,1.22 -6.12,3.06 -8.57,5.51c-3.06,2.45 -6.12,4.9 -9.18,6.73c-4.9,3.06 -0.61,10.4 4.9,7.96c6.73,-3.06 9.79,-11.02 15.91,-14.69C129.37,232.63 128.15,227.73 125.09,228.96z"/>
+    <path
+        android:fillColor="#D80027"
+        android:pathData="M85.92,257.11c-4.9,-0.61 -9.18,0 -14.08,1.22c-3.67,0.61 -7.96,1.22 -9.79,4.9c-1.22,1.84 0,4.9 2.45,5.51c3.67,0.61 6.12,-1.23 8.57,-3.06c3.67,-2.45 7.96,-4.9 12.85,-6.12C87.76,259.56 87.76,257.11 85.92,257.11z"/>
+</vector>
