+++ app/src/main/res/layout/activity_player_list_admin.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.pedrodavidlp.ittrivial.match.PlayerListAdminActivity">
+    tools:context="com.pedrodavidlp.ittrivial.join.PlayerListAdminActivity">
 
     <Button
         android:id="@+id/startGameButton"
